refactor: move update_checker from shared/ to adapters/config/
- Move version check service to adapters layer (infrastructure concern)
- Leave backward-compat shim in shared/
- Update lazy imports in cli.py and runtime_service.py
- Add to architecture test shim allowlist

269 tests pass, lint clean.
